refactor(dom): 로또 당첨 결과표 생성 로직을 util에서 controller로 이동
