+++ src/controllers/WebLottoController.js
@@ -1,6 +1,8 @@
 import CONFIG from '../constants/config.js';
+import PRIZE from '../constants/prize.js';
 import LotteryMachine from '../domain/services/LotteryMachine.js';
 import lottoService from '../domain/services/lottoService.js';
+import dom from '../utils/dom.js';
 import webOutputView from '../views/webOutputView.js';
 
 class WebLottoController {
@@ -24,7 +26,11 @@ class WebLottoController {
 
   #processLottoResult(matchedResultList) {
     const rankList = lottoService.calculateRankCounts(matchedResultList);
-    webOutputView.printLottoResult(rankList);
+
+    webOutputView.printLottoResult(
+      { createLottoResultTitle: this.createLottoResultTitle, createLottoResultTable: this.createLottoResultTable },
+      rankList,
+    );
   }
 
   #processProfit(matchedResultList) {
@@ -37,6 +43,32 @@ class WebLottoController {
     this.#processLottoResult(matchedResultList);
     this.#processProfit(matchedResultList);
   };
+
+  createLottoResultTitle() {
+    const lottoResultRowTitle = document.createElement('div');
+    lottoResultRowTitle.classList.add('lotto-result-row');
+    lottoResultRowTitle.classList.add('table-title');
+    PRIZE.TABLE_TITLE_LIST.forEach(text => {
+      const lottoTableTitle = dom.create('div', null, 'lotto-result-cell', text);
+      lottoResultRowTitle.appendChild(lottoTableTitle);
+    });
+    return lottoResultRowTitle;
+  }
+
+  createLottoResultTable(ranks) {
+    const resultTableFragment = document.createDocumentFragment();
+    [PRIZE.FIFTH, PRIZE.FORTH, PRIZE.THIRD, PRIZE.SECOND, PRIZE.FIRST].forEach(rank => {
+      const lottoResultRow = dom.create('div', null, 'lotto-result-row');
+      const lottoMatchedCount = dom.create('div', null, 'lotto-result-cell', `${PRIZE.COUNT_OUTPUTS[rank]}`);
+      const lottoResultPrize = dom.create('div', null, 'lotto-result-cell', `${PRIZE.AMOUNT[rank].toLocaleString()}`);
+      const lottoRankCount = dom.create('div', null, 'lotto-result-cell', `${ranks[rank]}개`);
+      lottoResultRow.appendChild(lottoMatchedCount);
+      lottoResultRow.appendChild(lottoResultPrize);
+      lottoResultRow.appendChild(lottoRankCount);
+      resultTableFragment.appendChild(lottoResultRow);
+    });
+    return resultTableFragment;
+  }
 }
 
 export default WebLottoController;
